Split benchmarks into per-type files, remove all regions
- Each benchmark class gets its own file (13 type files + InputKind.cs)
- Remove all #region/#endregion directives from generated code,
  tests, and NetworkData
- BenchmarkSwitcher.FromAssembly already supports --filter for CI
