+++ SortingNetworks.Benchmarks/CharSortingBenchmarks.cs
@@ -1,0 +1,48 @@
+using BenchmarkDotNet.Attributes;
+using SortingNetworks;
+
+namespace SortingNetworks.Benchmarks;
+
+[MemoryDiagnoser]
+[ShortRunJob]
+public class CharSortingBenchmarks
+{
+    private const int OpsPerInvoke = 1000;
+
+    [Params(27, 28)]
+    public int Length { get; set; }
+
+    private char[] _source = null!;
+    private char[][] _batch = null!;
+
+    [GlobalSetup]
+    public void Setup()
+    {
+        var rng = new Random(42);
+        _source = Enumerable.Range(0, Length).Select(_ => (char)rng.Next(32, 127)).ToArray();
+        _batch = new char[OpsPerInvoke][];
+        for (int i = 0; i < OpsPerInvoke; i++)
+            _batch[i] = new char[Length];
+    }
+
+    [IterationSetup]
+    public void IterationSetup()
+    {
+        for (int i = 0; i < OpsPerInvoke; i++)
+            Array.Copy(_source, _batch[i], Length);
+    }
+
+    [Benchmark(Baseline = true, OperationsPerInvoke = OpsPerInvoke)]
+    public void ArraySort()
+    {
+        for (int i = 0; i < OpsPerInvoke; i++)
+            Array.Sort(_batch[i]);
+    }
+
+    [Benchmark(OperationsPerInvoke = OpsPerInvoke)]
+    public void NetworkSort_Span()
+    {
+        for (int i = 0; i < OpsPerInvoke; i++)
+            NetworkSort.Sort(_batch[i].AsSpan());
+    }
+}
