+++ SortingNetworks.Benchmarks/InputKind.cs
@@ -1,0 +1,9 @@
+namespace SortingNetworks.Benchmarks;
+
+public enum InputKind
+{
+    Random,
+    Sorted,
+    Reversed,
+    Duplicates
+}
